Unify Robot and Task classes between training and simulation
